demo/requirements.txt
attrs==21.4.0
certifi==2024.8.30
cftime==1.6.4
charset-normalizer==2.0.12
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
dask==2024.10.0
debugpy==1.8.0
django==4.2.16
django-cors-headers==3.12.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
fiona==1.10.1
fsspec==2022.5.0
geographiclib==1.52
geopandas==0.14.2
geopy==2.2.0
idna==3.10
inflection==0.5.1
iniconfig==1.1.1
locket==1.0.0
netCDF4==1.7.0
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
partd==1.4.0
Pint==0.19.1
pluggy==1.0.0
psycopg2-binary==2.9.9
ptvsd==4.3.2
PyJWT==2.4.0
pyparsing==3.0.9
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0.1
requests==2.32.2
scipy==1.13
shapely==2.0.2
simple-file-checksum==1.1.0
six==1.16.0
sqlparse==0.5.1
tablib==3.6.1
tomli==2.0.1
toolz==0.12.1
uritemplate==4.1.1
urllib3==2.2.3
xarray==2024.9.0
XlsxWriter==3.0.3
